just commit

+++ src/ru/vsu/cs/valeev/test/PieDrawerImpl.java
@@ -16,9 +16,9 @@ public class PieDrawerImpl implements PieDrawer {
     public void drawPie(int x, int y, int width, int height, double startAngle, double arcAngle, Color c) {
         int centerX = x + width / 2;
         int centerY = y + height / 2;
-        int aStart = (int) (180 * startAngle / Math.PI);
-        int aArc = (int) (180 * arcAngle / Math.PI);
-        int aEnd = aStart + aArc;
-        bpd.draw(centerX, centerY, width / 2, height / 2, aStart, aEnd, c);
+        double aStart = 180 * startAngle / Math.PI;
+        double aArc = 180 * arcAngle/ Math.PI;
+        double aEnd = aStart + aArc;
+        bpd.draw(centerX, centerY, width / 2, height / 2, (int)aStart, (int)aEnd, c);
     }
 }
